+++ server/src/main/java/nbc/newsfeed/domain/dto/auth/request/TokenRefreshRequest.java
@@ -1,4 +1,4 @@
-package nbc.newsfeed.domain.dto.user.auth.request;
+package nbc.newsfeed.domain.dto.auth.request;
 
 import jakarta.validation.constraints.NotBlank;
